Arrange when go left to get through the ask_time limit line Previously when you select an event e.g. in single or in multi EPG but also when you go to primetime there is a hard line at the left from the GMEPG so you cannnot scroll left back to 'now'. This commit opens that barrier so you can always scroll left back to now.
Might still have some finetuning needed as it does not go always to the
previous event (but on the first one of the 'just' created new case. But
it is already better as it e.g. resolved the following scenario:

Open the GMEPG:

- select a random programnme
- Press the EPG INFO button
- Press Yellow to open the single EPG for the selected channel
- Scroll down a bit
- Press Exit

You return the the GMPEG, but you can't scroll back anymore to before the progamme you selected in the single EPG

https://forums.openpli.org/topic/16424-some-questions-bugs/page-353
Thanks to @Littlesat